# Speed up for accessing field of Struct.