@buruzaemon 2016-06-05
Thanks for the pull request, @pocke. I will check out your patch and consider adding this to a new release.